[refactor] 애플리케이션 계층 Exception 정리 #46
