+++ src/main/kotlin/dev/codeswamp/core/article/application/exception/application/AppBadRequestException.kt
@@ -1,0 +1,3 @@
+package dev.codeswamp.core.article.application.exception.application
+
+open class AppBadRequestException(message: String): AppException(message)
