+++ src/main/kotlin/dev/codeswamp/core/article/application/exception/application/AppConflictException.kt
@@ -1,0 +1,3 @@
+package dev.codeswamp.core.article.application.exception.application
+
+open class AppConflictException(message: String): AppException(message)
