+++ src/main/kotlin/dev/codeswamp/core/article/application/exception/application/AppException.kt
@@ -1,0 +1,4 @@
+package dev.codeswamp.core.article.application.exception.application
+
+open class AppException(message: String): RuntimeException(message)
+
