+++ src/main/kotlin/dev/codeswamp/core/article/application/exception/application/AppForbiddenException.kt
@@ -1,0 +1,3 @@
+package dev.codeswamp.core.article.application.exception.application
+
+open class AppForbiddenException(message: String): AppException(message)
