+++ src/main/kotlin/dev/codeswamp/core/article/application/exception/application/AppNotFoundException.kt
@@ -1,0 +1,3 @@
+package dev.codeswamp.core.article.application.exception.application
+
+open class AppNotFoundException(message: String): AppException(message)
