+++ src/main/kotlin/dev/codeswamp/core/article/application/rebase/RebasePolicy.kt
@@ -1,4 +1,4 @@
-package dev.codeswamp.core.article.application.base
+package dev.codeswamp.core.article.application.rebase
 
 import dev.codeswamp.core.article.domain.article.model.Version
