+++ src/main/kotlin/dev/codeswamp/core/article/application/usecase/command/publish/PublishArticleUseCaseImpl.kt
@@ -1,7 +1,7 @@
 package dev.codeswamp.core.article.application.usecase.command.publish
 
-import dev.codeswamp.core.article.application.base.RebasePolicy
-import dev.codeswamp.core.article.domain.article.exception.article.ArticleNotFoundException
+import dev.codeswamp.core.article.application.rebase.RebasePolicy
+import dev.codeswamp.core.article.application.exception.article.ArticleNotFoundException
 import dev.codeswamp.core.article.domain.article.model.VersionedArticle
 import dev.codeswamp.core.article.domain.article.model.vo.ArticleMetadata
 import dev.codeswamp.core.article.domain.article.model.vo.Slug
@@ -59,7 +59,7 @@ class PublishArticleUseCaseImpl(
     override fun update(command: UpdatePublishCommand) : PublishArticleResult {
         val createdAt = Instant.now()
 
-        val article = articleRepository.findByIdAndVersionId(command.articleId, command.versionId )
+        val article = articleRepository.findByIdAndVersionId(command.articleId, command.versionId)
             ?.apply { checkOwnership(command.userId) }
             ?.updateMetadata(ArticleMetadata(
                 folderId = command.folderId,
@@ -75,7 +75,7 @@ class PublishArticleUseCaseImpl(
                 rebasePolicy::shouldStoreAsBase,
                 contentReconstructor::reconstructFullContent)
             ?.publish(slugUniquenessChecker::checkSlugUniqueness)
-            ?: throw ArticleNotFoundException("Draft 저장에 실패했습니다 ")
+            ?: throw ArticleNotFoundException.byId(command.articleId)
 
         val saved = articleRepository.save(article)
         article.pullEvents().forEach(eventPublisher::publishEvent)
