+++ src/main/kotlin/dev/codeswamp/core/article/application/usecase/query/read/withversion/GetVersionedArticleUseCaseImpl.kt
@@ -1,7 +1,7 @@
 package dev.codeswamp.core.article.application.usecase.query.read.withversion
 
 import dev.codeswamp.core.article.application.usecase.query.read.ReadArticleResult
-import dev.codeswamp.core.article.domain.article.exception.article.ArticleNotFoundException
+import dev.codeswamp.core.article.application.exception.article.ArticleNotFoundException
 import dev.codeswamp.core.article.domain.article.repository.ArticleRepository
 import dev.codeswamp.core.article.domain.article.service.ArticleContentReconstructor
 import org.springframework.stereotype.Service
